+++ crates/biome_css_formatter/src/css/bogus/bogus_unicode_range_value.rs
@@ -1,0 +1,5 @@
+use crate::FormatBogusNodeRule;
+use biome_css_syntax::CssBogusUnicodeRangeValue;
+#[derive(Debug, Clone, Default)]
+pub(crate) struct FormatCssBogusUnicodeRangeValue;
+impl FormatBogusNodeRule<CssBogusUnicodeRangeValue> for FormatCssBogusUnicodeRangeValue {}