+++ crates/biome_css_parser/tests/spec_test.rs
@@ -174,10 +174,7 @@ pub fn run(test_case: &str, _snapshot_name: &str, test_directory: &str, outcome_
 #[test]
 pub fn quick_test() {
     let code = r#"
-div {
-  filter: progid:DXImageTransform.Microsoft.BasicImage(rotation=3);
-}
-
+@font-face { color: U+0100-024F; }
 
     "#;