+++ CHANGELOG.md
@@ -11,6 +11,12 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
 ## Unreleased
 
+### Parser
+
+#### New features
+
+- Implement [CSS unicode range](https://github.com/biomejs/biome/pull/3251). Contributed by @denbezrukov
+- 
 ### Formatter
 
 #### Bug fixes